Fix processing of double Content-DPR values and add Content-DPR WPT tests
When processing multiple Content-DPR headers, the current logic parses
all of them as a single float, and understandably, fails in that task.
That means that if a server is adding multiple headers (due to
configuration error), none of them will be effective, contrary to the RFC.

This PR fixes that, copies over relevant tests from internal layout tests
to WPT, and adds a test for that issue specifically.

BUG=895245

Change-Id: I72b570ad4b1e4db9a2bb03be9d8fb1e3799b902b